Update README.md
It seems tpc.org does not allow the TCP-DS zip file to be directly downloaded by the link in the Makefile any more.  The TPC-H zip file has a direct link to the .zip and I tried to construct a valid URL for the DS on without luck.  No joy on searching for an alternative either.